deleted all unneccessary files and folders
