+++ src/DefaultLayout/DefaultLayout.module.css
@@ -1,4 +1,4 @@
-@import '../../../index.module.css';
+@import '../index.module.css';
 
 .container {
 	display: grid;
