+++ src/component/APIFetch/APIFetch.js
@@ -1,12 +1,11 @@
-
 const APIFetch = () => {
-  return fetch("http://www.mocky.io/v2/5ea446a43000005900ce2ca3")
-    .then(function (response) {
-      return response.json();
-    })
-    .then(function (data) {
-      return data;
-    });
+	return fetch('http://www.mocky.io/v2/5ea446a43000005900ce2ca3')
+		.then(function(response) {
+			return response.json();
+		})
+		.then(function(data) {
+			return data;
+		});
 };
 
 export default APIFetch;
