+++ src/component/Backdrop/Backdrop.js
@@ -2,11 +2,7 @@ import React from 'react';
 import style from './Backdrop.module.css';
 
 const Backdrop = () => {
-    return (
-        <div className={style.backdrop}>
-
-        </div>
-    )
-}
+	return <div className={style.backdrop}></div>;
+};
 
 export default Backdrop;
