+++ src/component/Backdrop/Backdrop.module.css
@@ -1,5 +1,5 @@
 .backdrop {
-width: 100%;
-height: 100vh;
-background: rgba(0, 0, 0, 0.75);
+	width: 100%;
+	height: 100vh;
+	background: rgba(0, 0, 0, 0.75);
 }
