+++ src/component/Buttons/AddButton/AddButton.js
@@ -1,17 +1,12 @@
-import React, { useState } from 'react';
+import React from 'react';
 import style from './AddButton.module.css';
 
-
 const AddButton = () => {
-  return (
-        <div>
-    
-            <button className={style.buttonStyle}>+
-		        </button>
-
-        </div>
-    )
-
-}
+	return (
+		<div>
+			<button className={style.buttonStyle}>+</button>
+		</div>
+	);
+};
 
 export default AddButton;
