+++ src/component/Buttons/AddButton/AddButton.module.css
@@ -1,20 +1,18 @@
-
 .buttonStyle {
-    background: #0071B8;
-    border: 2px solid #FFFFFF;
-    box-sizing: border-box;
-    font-family: Roboto;
-    font-style: normal;
-    font-weight: normal;
-    font-size: 36px;
-    line-height: 42px;
-    display: flex;
-    align-items: center;
-    border-radius: 50%;
-    color: #FFFFFF;
-    width: 66.67px;
-    height: 66.67px;
-    justify-content: center;
-    margin: 0 28.66px 28.66px 28.66px;
-    
+	background: #0071b8;
+	border: 2px solid #ffffff;
+	box-sizing: border-box;
+	font-family: Roboto;
+	font-style: normal;
+	font-weight: normal;
+	font-size: 36px;
+	line-height: 42px;
+	display: flex;
+	align-items: center;
+	border-radius: 50%;
+	color: #ffffff;
+	width: 66.67px;
+	height: 66.67px;
+	justify-content: center;
+	margin: 0 28.66px 28.66px 28.66px;
 }
